requirements.txt
django==5.0.1
django-admin-interface==0.19.1
django-admin-honeypot==1.1.0
django-ipware==4.0.2
python-dotenv==0.20.0
sentry-sdk==1.6.0
django-extra-settings
django-allauth==0.51.0
django-hCaptcha==0.2.0
whitenoise==6.2.0
django-maintenance-mode==0.16.3
# django-debug-toolbar==3.5.0
django-pwa==1.0.10
django-admin-interface==0.19.1
pyminizip
discord-webhook
pytz
djlint
django-extensions
Werkzeug
